# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CCP IONEX — Clinic Control Panel API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clinics/{clinicId}/team:
    get:
      operationId: listTeam
      tags: [team]
      summary: List internal team members for a clinic
      parameters:
        - name: clinicId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Team members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeamMember"
    post:
      operationId: createTeamMember
      tags: [team]
      summary: Add a team member
      parameters:
        - name: clinicId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTeamMemberBody"
      responses:
        "201":
          description: Created team member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"
  /team/{id}:
    patch:
      operationId: updateTeamMember
      tags: [team]
      summary: Update a team member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTeamMemberBody"
      responses:
        "200":
          description: Updated team member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"
    delete:
      operationId: deleteTeamMember
      tags: [team]
      summary: Delete a team member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ─── FATURAS ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    TeamMember:
      type: object
      properties:
        id:
          type: string
        clinicId:
          type: string
        nome:
          type: string
        funcao:
          type: ["string", "null"]
        area:
          type: ["string", "null"]
        vinculo:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        whatsapp:
          type: ["string", "null"]
        temAcessoPlataforma:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - clinicId
        - nome
        - temAcessoPlataforma
        - createdAt
    CreateTeamMemberBody:
      type: object
      properties:
        nome:
          type: string
        funcao:
          type: ["string", "null"]
        area:
          type: ["string", "null"]
        vinculo:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        whatsapp:
          type: ["string", "null"]
        temAcessoPlataforma:
          type: boolean
      required:
        - nome
    UpdateTeamMemberBody:
      type: object
      properties:
        nome:
          type: ["string", "null"]
        funcao:
          type: ["string", "null"]
        area:
          type: ["string", "null"]
        vinculo:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        whatsapp:
          type: ["string", "null"]
        temAcessoPlataforma:
          type: ["boolean", "null"]

    # ─── FATURA ─────────────────────────────────────────────────────────────
...